+++ cmd/uymas/mian.go
@@ -202,19 +202,29 @@ func (c *defaultApp) Scan() {
 		baseDir = "./"
 	}
 	dd := fs.NewDirScanner(baseDir)
+	dd.CddChanMax = cc.ArgInt("bufsize", "B")
 
 	//过滤
 	dd.Exclude(cc.ArgStringSlice("exclude")...)
 	dd.Include(cc.ArgStringSlice("include")...)
 
-	if er := dd.Scan(); er == nil {
+	var er error
+	var isParallel = "否"
+	if cc.CheckSetting("parallel", "ll") {
+		er = dd.ScanParallel()
+		isParallel = "是"
+	} else {
+		er = dd.Scan()
+	}
+
+	if er == nil {
 		var table = [][]interface{}{{"Path", "Size", "Depth"}}
 		for key, tcd := range dd.TopChildDick {
 			table = append(table, []interface{}{key, number.Bytes(tcd.Size), tcd.Depth})
 		}
 
 		fmt.Println(bin.FormatTable(table, false))
-		fmt.Printf(" 文件扫目标目录： %v.\r\n", dd.BaseDir())
+		fmt.Printf(" 文件扫目标目录： %v，是否并发: %v(线程分配 %v).\r\n", dd.BaseDir(), isParallel, dd.ChanNumber())
 		fmt.Printf(" 文件扫描数： %v, 目录: %v, 文件： %v.\r\n", dd.AllItem, dd.AllDirItem, dd.AllFileItem)
 		fmt.Printf(" 目录大小: %v.\r\n", number.Bytes(dd.AllSize))
 		fmt.Printf(" 使用时间： %v.\r\n", dd.Runtime)